pyproject.toml
[tool.poetry]
name = "weather-app"
version = "0.1.0"
description = ""
authors = ["fed0r <wisp47344@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.8"
sqlalchemy = "^2.0.38"
alembic = "^1.14.1"
dependency-injector = "^4.46.0"
python-dotenv = "^1.0.1"
asyncpg = "^0.30.0"
aiohttp = "^3.11.13"
uvicorn = "^0.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
